Abstractized the in-memory store.
