+++ package.json
@@ -51,7 +51,6 @@
     "mqtt-packet": "^3.2.0",
     "readable-stream": "~1.0.2",
     "websocket-stream": "^1.3.1",
-    "ws": "^0.6.3",
     "xtend": "^4.0.0"
   },
   "devDependencies": {
@@ -65,6 +64,7 @@
     "sinon": "~1.10.0",
     "through2": "^0.6.3",
     "uglify": "^0.1.1",
+    "ws": "^0.7.1",
     "zuul": "^1.15.2"
   }
 }
